+++ pollbot/pollbot.py
@@ -66,9 +66,10 @@ dispatcher.add_handler(InlineQueryHandler(search))
 dispatcher.add_handler(ChosenInlineResultHandler(handle_chosen_inline_result))
 
 
+minute = 60
 job_queue = updater.job_queue
 job_queue.run_repeating(message_update_job, interval=1, first=0, name='Handle poll message update queue')
-job_queue.run_repeating(send_notifications, interval=10, first=0, name='Handle notifications and due dates')
+job_queue.run_repeating(send_notifications, interval=5*minute, first=0, name='Handle notifications and due dates')
 
 # Message handler
 dispatcher.add_handler(